# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Music Player API
  description: API for managing music playlists (groups), tracks, and playback board
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  parameters:
      TrackId:
        name: trackId
        in: path
        description: ID of the track
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      PointId:
        name: pointId
        in: path
        description: ID of the point
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      UserId:
        name: userId
        in: path
        description: ID of the user
        required: true
        schema:
          type: integer
          format: int64
          example: 1
...
  schemas:
    User:
      type: object
      description: Basic user representation
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the user
          example: 1
        name:
          type: string
          description: Unique username
          example: "bruno"
...
    TrackShareRequest:
      type: object
      description: Request to share a track with another user
      required:
        - userId
      properties:
        userId:
          type: integer
          format: int64
          description: ID of the user to share the track with
          example: 2
    TrackShare:
      type: object
      description: Track sharing information for a specific user
      properties:
        user:
          $ref: '#/components/schemas/User'
...
    Track:
      type: object
      description: A music track
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier (auto-generated)
          example: 1
        trackName:
          type: string
          description: Name or title of the track
          example: "Sample Song"
        trackLink:
          type: string
          description: URL or link to the track audio
          example: "https://example.com/track.mp3"
        duration:
          type: integer
          description: Duration of the track in seconds
          example: 215
        ownerId:
          type: integer
          format: int64
          description: ID of the user who owns this track
          example: 1
        groupIds:
          type: array
          description: IDs of groups this track belongs to
          items:
            type: integer
            format: int64
          example: [3, 7, 12]
        trackPoints:
          type: array
          description: Points of interest for the track
          items:
            $ref: '#/components/schemas/TrackPoint'
    TrackPoint:
      type: object
      description: Point of interest in track
      properties:
          id:
            type: integer
            format: int64
            description: Unique identifier (auto-generated)
          name:
              type: string
              description: name for point
          position:
            type: integer
            description: Point of interest in the track
          fadeIn:
            type: boolean
            description: point makes a fade in
          fadeOut:
            type: boolean
            description: point makes a fade out
    TrackRequest:
      type: object
      description: New track request
      required:
        - trackLink
      properties:
        trackName:
          type: string
          description: Name or title of the track. If not set, default is used
          example: "Sample Song"
        trackLink:
          type: string
          description: URL or link to the track audio
          example: "https://example.com/track.mp3"
    TrackPointRequest:
      type: object
      description: Point of interest in track
      properties:
        position:
          type: integer
          description: Point of interest in the track
        name:
          type: string
          description: name for point
        fadeIn:
          type: boolean
          description: point makes a fade in
        fadeOut:
          type: boolean
          description: point makes a fade out
...
paths:
...
  /tracks:
    post:
      tags:
        - MusicTracks
      summary: Create a new track
      operationId: createTrack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackRequest'
      responses:
        '201':
          description: Track created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Invalid input
        '409':
          description: Conflict (e.g., duplicate track link)
    get:
      tags:
        - MusicTracks
      summary: Get list of tracks for a user
      operationId: getUserTracks
      responses:
        '200':
          description: List of tracks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '404':
          description: User not found
  /tracks/{trackId}:
    put:
      tags:
        - MusicTracks
      summary: Update an existing track
      operationId: updateTrack
      parameters:
          - $ref: '#/components/parameters/TrackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackRequest'
      responses:
        '200':
          description: Track updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Invalid input
        '404':
          description: Track not found
    delete:
      tags:
        - MusicTracks
      summary: Delete a track
      operationId: deleteTrack
      parameters:
          - $ref: '#/components/parameters/TrackId'
      responses:
        '204':
          description: Track deleted successfully
        '404':
          description: Track not found
  /tracks/{trackId}/points:
    post:
      tags:
         - MusicTracks
      summary: Create a track point
      description: Adds a new point of interest to the specified track
      operationId: createTrackPoint
      parameters:
          - $ref: '#/components/parameters/TrackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackPointRequest'
      responses:
        '200':
          description: Updated track with the new point
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request (validation error, etc.)
        '500':
          description: Server error
  /tracks/{trackId}/points/{pointId}:
    patch:
      tags:
        - MusicTracks
      summary: Update a track point
      description: Partially updates the specified point of interest on the track
      operationId: updateTrackPoint
      parameters:
        - $ref: '#/components/parameters/TrackId'
        - $ref: '#/components/parameters/PointId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackPointRequest'
      responses:
        '200':
          description: Updated track with the modified point
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request (validation error, not found, etc.)
        '500':
          description: Server error
    delete:
      tags:
        - MusicTracks
      summary: Delete a track point
      description: Removes the specified point from the track
      operationId: deleteTrackPoint
      parameters:
        - $ref: '#/components/parameters/TrackId'
        - $ref: '#/components/parameters/PointId'
      responses:
        '200':
          description: Updated track without the point
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request (not found, etc.)
        '500':
          description: Server error
  /tracks/{trackId}/shares:
    get:
      tags:
        - MusicTracks
      summary: List users the track is shared with
      description: Returns all users that have view access to this track. Only the owner can call this endpoint.
      operationId: listTrackShares
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: List of users the track is shared with
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackShare'
        '403':
          description: Forbidden (only the owner can view shares)
        '404':
          description: Track not found
    post:
      tags:
        - MusicTracks
      summary: Share a track with a user
      description: >
        Shares the track with another user. If already shared, this is a no-op.
        Only the owner can call this endpoint.
      operationId: shareTrack
      parameters:
        - $ref: '#/components/parameters/TrackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackShareRequest'
      responses:
        '200':
          description: Track sharing created (or already existed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackShare'
        '400':
          description: Invalid input
        '403':
          description: Forbidden (only the owner can share)
        '404':
          description: Track or user not found
  /tracks/{trackId}/shares/{userId}:
    delete:
      tags:
        - MusicTracks
      summary: Remove sharing of a track for a user
      description: >
        Removes access to the track for the specified user.
        Only the owner can call this endpoint.
      operationId: unshareTrack
      parameters:
        - $ref: '#/components/parameters/TrackId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Sharing removed successfully
        '403':
          description: Forbidden (only the owner can unshare)
        '404':
          description: Track or sharing not found
...